ToDo list by JS
Previously, I always ignored a project like TODO LIST, as it seemed to me that it was too simple. However, I was wrong and this turned out to be a great practice for improving skills in manipulating the DOM tree.
Also, I created the functions for adding, deleting and changing elements in localStorage for access after reloading the web page to active tasks.
And to top it off, "LiveSearch" has been added to the page to sort the tasks.